no unecessary encoders
