+++ packages/common/src/services/decision/reviewHelpers.ts
@@ -1,10 +1,9 @@
 import { db } from '@op/db/client';
 import type { User } from '@op/supabase/lib';
 
-import { NotFoundError, UnauthorizedError, ValidationError } from '../../utils';
+import { NotFoundError, UnauthorizedError } from '../../utils';
 import { assertUserByAuthId } from '../assert';
 import { getInstance } from './getInstance';
-import type { RubricTemplateSchema } from './types';
 
 /** Loads and authorizes access to a single review assignment for the current reviewer. */
 export async function getAuthorizedReviewAssignmentContext({
@@ -14,7 +13,7 @@ export async function getAuthorizedReviewAssignmentContext({
   assignmentId: string;
   user: User;
 }) {
-  const [assignment, commonUser] = await Promise.all([
+  const [assignment, dbUser] = await Promise.all([
     db.query.proposalReviewAssignments.findFirst({
       where: {
         id: assignmentId,
@@ -30,7 +29,7 @@ export async function getAuthorizedReviewAssignmentContext({
     throw new NotFoundError('Review assignment');
   }
 
-  if (!commonUser.profileId) {
+  if (!dbUser.profileId) {
     throw new UnauthorizedError('User must have an active profile');
   }
 
@@ -39,40 +38,21 @@ export async function getAuthorizedReviewAssignmentContext({
     user,
   });
 
+  // TODO: revisit the access here
   if (!instance.access.review && !instance.access.admin) {
     throw new UnauthorizedError("You don't have access to review proposals");
   }
 
-  if (assignment.reviewerProfileId !== commonUser.profileId) {
+  if (assignment.reviewerProfileId !== dbUser.profileId) {
     throw new UnauthorizedError(
       "You don't have access to this review assignment",
     );
   }
 
-  const instanceData =
-    instance.instanceData && typeof instance.instanceData === 'object'
-      ? instance.instanceData
-      : null;
-
   return {
     assignment,
     instance,
-    review: assignment.reviews[0] ?? null,
-    rubricTemplate:
-      (instanceData?.rubricTemplate as RubricTemplateSchema | undefined) ??
-      null,
+    review: assignment.reviews[0],
+    rubricTemplate: instance.instanceData.rubricTemplate,
   };
-}
-
-/** Returns the rubric template required to save or submit a review. */
-export function requireRubricTemplate(
-  rubricTemplate: RubricTemplateSchema | null,
-): RubricTemplateSchema {
-  if (!rubricTemplate) {
-    throw new ValidationError(
-      'Review rubric is not configured for this process',
-    );
-  }
-
-  return rubricTemplate;
 }
