+++ packages/common/src/services/decision/schemas/index.ts
@@ -3,4 +3,3 @@ export * from './definitions';
 export * from './instanceData';
 export * from './reviews';
 export * from './proposal';
-export * from './reviewAssignments';
