+++ packages/common/src/services/decision/schemas/reviews.ts
@@ -1,9 +1,30 @@
-import { ProposalReviewState } from '@op/db/schema';
+import {
+  ProposalReviewAssignmentStatus,
+  ProposalReviewState,
+} from '@op/db/schema';
 import { z } from 'zod';
 
-import { proposalReviewAssignmentListItemSchema } from './reviewAssignments';
+import { proposalSchema } from './proposal';
 
 const jsonObjectSchema = z.record(z.string(), z.unknown());
+
+// ── Review assignment schemas ───────────────────────────────────────────
+
+export const proposalReviewAssignmentListItemSchema = z.object({
+  id: z.uuid(),
+  processInstanceId: z.uuid(),
+  phaseId: z.string(),
+  status: z.nativeEnum(ProposalReviewAssignmentStatus),
+  proposal: proposalSchema,
+});
+
+export const proposalReviewAssignmentListSchema = z.object({
+  assignments: z.array(proposalReviewAssignmentListItemSchema),
+  total: z.number(),
+  completed: z.number(),
+});
+
+// ── Review schemas ──────────────────────────────────────────────────────
 
 export const proposalReviewSchema = z.object({
   id: z.uuid(),
@@ -22,6 +43,14 @@ export const reviewAssignmentDetailSchema = z.object({
   review: proposalReviewSchema.nullable(),
 });
 
+// ── Types ───────────────────────────────────────────────────────────────
+
+export type ProposalReviewAssignmentListItem = z.infer<
+  typeof proposalReviewAssignmentListItemSchema
+>;
+export type ListProposalReviewAssignmentsResult = z.infer<
+  typeof proposalReviewAssignmentListSchema
+>;
 export type ProposalReviewData = z.infer<typeof proposalReviewSchema>;
 export type ProposalReview = z.infer<typeof proposalReviewSchema>;
 export type ReviewAssignmentDetail = z.infer<
